Cast the fixed point value to float explicitly.

+++ python/fusion_engine_client/utils/construct_utils.py
@@ -23,7 +23,7 @@ class FixedPointAdapter(Adapter):
         if obj == self.invalid:
             return math.nan
         else:
-            return obj * self.scale
+            return float(obj * self.scale)
 
     def _encode(self, obj, context, path):
         if math.isnan(obj) and self.invalid is not None:
